requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.1.0
astroid==3.0.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.3
dill==0.3.7
distro==1.8.0
exceptiongroup==1.2.0
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.12.2
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
inflection==0.5.1
isort==5.13.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
langchain==0.0.350
langchain-community==0.0.3
langchain-core==0.1.1
langsmith==0.0.70
marshmallow==3.20.1
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.4.0
openapi==1.1.0
packaging==23.2
platformdirs==4.1.0
pydantic==2.5.2
pydantic_core==2.14.5
pylint==3.0.3
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
rpds-py==0.13.2
safetensors==0.4.1
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
tokenizers==0.15.0
tomli==2.0.1
tomlkit==0.12.3
tqdm==4.66.1
transformers==4.36.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
